Configurate Webpack

+++ src/controllers/videoController.js
@@ -17,9 +17,15 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   return res.render('edit', { pageTitle: `Edit ${video.title}`, video });
 };
@@ -27,9 +33,15 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -75,6 +87,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect('/');
 };
